Moves path of and references to ray task decorator
