+++ ray_provider/example_dags/task_flow_xgboost_modin.py
@@ -5,7 +5,7 @@ from airflow.decorators import dag, task
 from airflow.utils.dates import days_ago
 from airflow.operators.dummy_operator import DummyOperator
 import ray
-from ray_provider.operators.ray_decorators import ray_task
+from ray_provider.decorators.ray_decorators import ray_task
 import numpy as np
 import xgboost_ray as xgb
 from ray_provider.xcom.ray_backend import RayBackend
