+++ ray_provider/example_dags/xgboost_pandas_tune_breast_cancer.py
@@ -4,7 +4,7 @@ from airflow.decorators import dag, task
 from airflow.utils.dates import days_ago
 from airflow.operators.dummy_operator import DummyOperator
 import ray
-from ray_provider.operators.ray_decorators import ray_task
+from ray_provider.decorators.ray_decorators import ray_task
 import numpy as np
 import xgboost_ray as xgbr
 import xgboost as xgb
